Fix compilation on 32-bit host systems: explicit cast to ulong for %lu formats.
git-svn-id: https://freenos.googlecode.com/svn/trunk@226 824000e8-63cf-11de-a6e1-e33d94ba4d19

+++ srv/filesystem/linn/LinnCreate.h
@@ -57,7 +57,7 @@
         if ((sb)->freeBlocksCount < (count)) \
 	{ \
 	    printf("%s: not enough free blocks remaining (%lu needed)\n", \
-		    prog, (count)); \
+		    prog, (ulong)(count)); \
 	    exit(EXIT_FAILURE); \
 	} \
 	(sb)->freeBlocksCount -= (count); \
